backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
autopep8==2.0.1
Django==3.2.7
django-autoslug==1.9.8
django-countries==7.2.1
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
inflection==0.5.1
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
phonenumbers==8.12.33
Pillow==9.4.0
psycopg2==2.9.6
pycodestyle==2.11.0
PyJWT==2.1.0
python-dotenv==0.21.0
python-http-client==3.3.7
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.9.2
sendgrid==6.10.0
shortuuid==1.0.11
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
